--- src/builder/VsCodeBuilder.ts
@@ -1,20 +1,0 @@
-import { TypeScriptProject } from 'projen/lib/typescript';
-import constants from '../constants/templates';
-
-/**
- * Sets up the vscode config.
- */
-export class VsCodeBuilder {
-  private project: TypeScriptProject;
-
-  constructor(project: TypeScriptProject) {
-    this.project = project;
-  }
-
-  /**
-   * Adds the VSCode settings for the project.
-   */
-  public addSettings(): void {
-    this.project.vscode?.settings.addSettings(constants.VS_CODE.SETTINGS);
-  }
-}
