+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './NpmPackageComponent';
+export * from './DevContainerComponent';
+export * from './VsCodeComponent';
+export * from './PrettierComponent';
