+++ src/index.ts
@@ -1,6 +1,2 @@
 export * from './GitHubActionProject';
-export * from './builder/NpmPackageBuilder';
-export * from './builder/GitHubBuilder';
-export * from './builder/DevContainerBuilder';
-export * from './builder/VsCodeBuilder';
-export * from './builder/PrettierBuilder';
+export * from './components';
