+++ src/types/component.ts
@@ -1,0 +1,7 @@
+/**
+ * Defines methods for a component to be created in the project.
+ */
+export interface IProjectComponent {
+  add?(): void;
+  addScripts?(): void;
+}
